+++ load.go
@@ -11,16 +11,17 @@ import (
 type UnmarshalOptions = file.UnmarshalOptions
 
 // Load creates a uconfig manager with defaults,environment variables,
-// and optionally file loaders based on the provided
-// Files map.
-func Load[C any](files Files, userPlugins ...plugins.Plugin) Config[C] {
-	ps := make([]plugins.Plugin, 0, len(files)+2+len(userPlugins))
+// and optionally file loaders based on the provided PluginProvider.
+func Load[C any](files PluginProvider, userPlugins ...plugins.Plugin) Config[C] {
+	ps := make([]plugins.Plugin, 0, 2+len(userPlugins))
 
 	// first defaults
 	ps = append(ps, defaults.New())
 	// then files
-	ps = append(ps, files.Plugins()...)
-	// then any user pugins, often just _secret_.
+	if files != nil {
+		ps = append(ps, files.Plugins()...)
+	}
+	// then any user plugins, often just _secret_.
 	ps = append(ps, userPlugins...)
 
 	// followed by envs